src/performance_test.py
"""
English Version - Translated for international release
Date: 2026-02-27
Translator: AetherClaw Night Market Intelligence
"""
#!/usr/bin/env python3
import json
import orjson
import ujson
import rapidjson
import time
import sys
def test_json_performance():
    print("🧪 JSON")
    print("=" * 60)
    # Testing
    test_data = {
        "": {
            "": "v3.0-full",
            "": time.strftime("%Y-%m-%d %H:%M:%S"),
            "": "",
            "": {
                "items": [{"id": i, "name": f"{i}", "value": i * 10} for i in range(1000)],
                "metadata": {"": "AetherClaw", "": ""}
            }
        }
    }
    # Testing
    print("\n📊 :")
    print("-" * 40)
    results = {}
    # orjson
    start = time.perf_counter()
    for _ in range(100):
        orjson.dumps(test_data)
    results['orjson'] = (time.perf_counter() - start) * 1000 / 100
    # ujson
    start = time.perf_counter()
    for _ in range(100):
        ujson.dumps(test_data)
    results['ujson'] = (time.perf_counter() - start) * 1000 / 100
    # rapidjson
    start = time.perf_counter()
    for _ in range(100):
        rapidjson.dumps(test_data)
    results['rapidjson'] = (time.perf_counter() - start) * 1000 / 100
    # 
    start = time.perf_counter()
    for _ in range(100):
        json.dumps(test_data)
    results['stdlib'] = (time.perf_counter() - start) * 1000 / 100
    # 
    for lib, time_ms in sorted(results.items(), key=lambda x: x[1]):
        speedup = results['stdlib'] / time_ms if time_ms > 0 else 0
        print(f"  {lib:10s}: {time_ms:.3f}ms ({speedup:.1f}x)")
    # Testing
    print("\n📊 :")
    print("-" * 40)
    json_str = json.dumps(test_data)
    parse_results = {}
    # orjson
    start = time.perf_counter()
    for _ in range(100):
        orjson.loads(json_str.encode('utf-8'))
    parse_results['orjson'] = (time.perf_counter() - start) * 1000 / 100
    # ujson
    start = time.perf_counter()
    for _ in range(100):
        ujson.loads(json_str)
    parse_results['ujson'] = (time.perf_counter() - start) * 1000 / 100
    # rapidjson
    start = time.perf_counter()
    for _ in range(100):
        rapidjson.loads(json_str)
    parse_results['rapidjson'] = (time.perf_counter() - start) * 1000 / 100
    # 
    start = time.perf_counter()
    for _ in range(100):
        json.loads(json_str)
    parse_results['stdlib'] = (time.perf_counter() - start) * 1000 / 100
    # 
    for lib, time_ms in sorted(parse_results.items(), key=lambda x: x[1]):
        speedup = parse_results['stdlib'] / time_ms if time_ms > 0 else 0
        print(f"  {lib:10s}: {time_ms:.3f}ms ({speedup:.1f}x)")
    # Performance
    print("\n🎯 :")
    print("-" * 40)
    best_serialize = min(results, key=results.get)
    best_parse = min(parse_results, key=parse_results.get)
    print(f"  : {best_serialize} ({results[best_serialize]:.3f}ms)")
    print(f"  : {best_parse} ({parse_results[best_parse]:.3f}ms)")
    # XML
    xml_baseline = 100  # XML100ms
    json_performance = results[best_serialize] + parse_results[best_parse]
    speedup_vs_xml = xml_baseline / json_performance if json_performance > 0 else 0
    print(f"\n⚡ XML:")
    print(f"  XML: {xml_baseline}ms")
    print(f"  JSON: {json_performance:.1f}ms")
    print(f"  : {speedup_vs_xml:.1f}x ({(speedup_vs_xml-1)*100:.0f}%)")
    print("\n" + "=" * 60)
    print("✅ JSON")
    return {
        "serialize_results": results,
        "parse_results": parse_results,
        "best_serialize": best_serialize,
        "best_parse": best_parse,
        "speedup_vs_xml": speedup_vs_xml
    }
if __name__ == "__main__":
    test_json_performance()